Fixed warning: file.gcount() returns a negative integer.
So the compiler complains that the comparison ends in a change of
signedness, which can change the value.

+++ src/bin_file_reader.cc
@@ -27,7 +27,7 @@ static bool is_header_correct(std::fstream& file, const char expected[4])
   char buffer[4];
   file.read(buffer, sizeof(buffer));
 
-  return (file.gcount() == sizeof(buffer)) and
+  return (static_cast<unsigned>(file.gcount()) == sizeof(buffer)) and
     (std::memcmp(buffer, expected, sizeof(buffer)) == 0);
 }
